Believe I've got ownership split out properly, finally.
